requirements-windows.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
python-dotenv==1.0.0
requests==2.31.0
gspread==5.12.1
oauth2client==4.1.3
schedule==1.2.0
pytz==2023.3
pydantic==2.5.0
jinja2==3.1.2
aiofiles==23.2.1
python-multipart==0.0.6
aiohttp==3.9.1